# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "version": "1.0"
    },
    "paths": {
        "/api/passengers": {
            "get": {
                "summary": "Get All Passengers",
                "description": "Fetches all passengers with their associated rides",
                "operationId": "b82b3733013ad17a723c84b8dcca2f9b",
                "responses": {
                    "200": {
                        "description": "A list of passengers"
                    },
                    "404": {
                        "description": "No passengers found"
                    }
                }
            },
            "delete": {
                "summary": "Delete multiple passengers",
                "description": "Deletes multiple passengers by their IDs",
                "operationId": "8461bf59b3efa9f4bd957a832fac3f7e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "passenger_ids"
                                ],
                                "properties": {
                                    "passenger_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Passengers successfully deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "2 passengers deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No passengers found to delete"
                    }
                }
            }
        },
        "/api/passengers/{passengerID}": {
            "get": {
                "summary": "Get a specific passenger",
                "description": "Fetches a passenger by ID along with their associated rides",
                "operationId": "84ea540ac35cdd8bd61e0faa9ffc3710",
                "parameters": [
                    {
                        "name": "passengerID",
                        "in": "path",
                        "description": "ID of the passenger",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the specific passenger"
                    },
                    "404": {
                        "description": "Passenger not found"
                    }
                }
            }
        }
    }
}